Intake and ShootOnTheMoveFix
