+++ src/main/java/frc/robot/subsystems/IntakeRoller.java
@@ -4,15 +4,17 @@ import com.revrobotics.PersistMode;
 import com.revrobotics.ResetMode;
 import com.revrobotics.spark.SparkMax;
 import com.revrobotics.spark.config.SparkMaxConfig;
+import com.revrobotics.spark.config.SparkFlexConfig;
+import com.revrobotics.spark.SparkFlex;
 import frc.robot.Constants;
 import frc.robot.Constants.JamProtectionConstants;
 import frc.robot.Constants.MotorConstants;
 import frc.robot.util.JamProtection;
 import frc.robot.util.TunableNumber;
 
-public class IntakeRoller extends MaxActuator {
-  private SparkMax motor;
-  private SparkMaxConfig motorConfig;
+public class IntakeRoller extends FlexActuator {
+  private SparkFlex motor;
+  private SparkFlexConfig motorConfig;
   private static IntakeRoller instance;
 
   // Tunable operational value
@@ -48,7 +50,7 @@ public class IntakeRoller extends MaxActuator {
         false);
     motor = getMotor();
 
-    motorConfig = new SparkMaxConfig();
+    motorConfig = new SparkFlexConfig();
 
     motorConfig.idleMode(SparkMaxConfig.IdleMode.kCoast).smartCurrentLimit(40);
